requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
asyncpg>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.9.0
